--- runner/src/server/plugins/engine/pageControllers/homePageController.ts
@@ -1,28 +1,0 @@
-import { PageController } from "./pageController";
-import { HapiRequest, HapiResponseToolkit } from "server/types";
-
-export class HomePageController extends PageController {
-  get getRouteOptions() {
-    return {
-      ext: {
-        onPostHandler: {
-          method: (_request: HapiRequest, h: HapiResponseToolkit) => {
-            return h.continue;
-          },
-        },
-      },
-    };
-  }
-
-  get postRouteOptions() {
-    return {
-      ext: {
-        onPostHandler: {
-          method: (_request: HapiRequest, h: HapiResponseToolkit) => {
-            return h.continue;
-          },
-        },
-      },
-    };
-  }
-}
