--- runner/src/server/plugins/engine/pageControllers/pageController.ts
@@ -1,50 +1,0 @@
-import { PageControllerBase } from "./pageControllerBase";
-import { HapiRequest, HapiResponseToolkit } from "server/types";
-import { FormModel } from "../models";
-
-export class PageController extends PageControllerBase {
-  constructor(model: FormModel, pageDef: any) {
-    super(model, pageDef);
-  }
-
-  get getRouteOptions(): {
-    ext: any;
-  } {
-    return {
-      ext: {
-        onPostHandler: {
-          method: (_request: HapiRequest, h: HapiResponseToolkit) => {
-            return h.continue;
-          },
-        },
-      },
-    };
-  }
-
-  get postRouteOptions(): {
-    payload?: any;
-    ext: any;
-  } {
-    return {
-      payload: {
-        output: "stream",
-        parse: true,
-        maxBytes: Number.MAX_SAFE_INTEGER,
-        failAction: "ignore",
-      },
-      ext: {
-        onPreHandler: {
-          method: async (request: HapiRequest, h: HapiResponseToolkit) => {
-            const { uploadService } = request.services([]);
-            return uploadService.handleUploadRequest(request, h);
-          },
-        },
-        onPostHandler: {
-          method: async (_request: HapiRequest, h: HapiResponseToolkit) => {
-            return h.continue;
-          },
-        },
-      },
-    };
-  }
-}
